requirements.txt
asyncio~=3.4.3
dataclasses~=0.6
python-dotenv==0.21.0
APScheduler==3.9.1
mysql-connector-python==8.0.29
httpx==0.23.0
undetected-chromedriver==3.1.6
bs4~=0.0.1
beautifulsoup4==4.11.1
python-telegram-bot==13.14
